+++ frontend/src/App.css
@@ -1,490 +1,4 @@
-* {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  font-family: "Poppins", Arial, sans-serif;
-  background: #f8fafc;
-  color: #0f172a;
-}
-
-button {
-  font-family: inherit;
-}
-
 .app {
   min-height: 100vh;
   display: flex;
-}
-
-.sidebar {
-  width: 180px;
-  min-height: 100vh;
-  background: #15224a;
-  color: white;
-  padding: 28px 18px;
-  display: flex;
-  flex-direction: column;
-  position: sticky;
-  top: 0;
-}
-
-.logo-container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  padding-top: 28px;
-  margin-bottom: 40px;
-}
-
-.logo-image {
-  max-width: 170px;
-  height: auto;
-  object-fit: contain;
-  display: block;
-}
-
-.nav {
-  width: 100%;
-  border: 0;
-  color: white;
-  background: transparent;
-  padding: 14px 16px;
-  border-radius: 10px;
-  text-align: left;
-  cursor: pointer;
-  margin-bottom: 8px;
-  font-size: 14px;
-}
-
-.nav.active,
-.nav:hover {
-  background: #005aff;
-}
-
-.page {
-  flex: 1;
-  padding: 28px 32px;
-}
-
-.topbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 24px;
-}
-
-.topbar h1 {
-  margin: 0;
-  font-size: 28px;
-}
-
-.dashboard-grid {
-  display: grid;
-  grid-template-columns: 1.15fr 1.35fr;
-  gap: 24px;
-  margin-bottom: 24px;
-}
-
-.dashboard-map-panel,
-.dashboard-kpi-panel,
-.map-card,
-.kpi-card,
-.last-session,
-.phone-list,
-.phone-detail,
-.summary-row,
-.recent-sessions,
-.parameter-empty,
-.empty-box {
-  background: white;
-  border: 1px solid #dbe3ef;
-  border-radius: 14px;
-  box-shadow: 0 12px 30px rgba(15, 23, 42, 0.04);
-}
-
-.dashboard-map-panel,
-.dashboard-kpi-panel,
-.map-card,
-.last-session,
-.phone-list,
-.phone-detail,
-.recent-sessions,
-.parameter-empty,
-.empty-box {
-  padding: 18px;
-}
-
-.dashboard-kpi-panel {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-h2,
-h3 {
-  margin: 0 0 12px;
-}
-
-.tabs {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 8px;
-  margin-bottom: 14px;
-}
-
-.tab {
-  border: 1px solid #dbe3ef;
-  background: #f8fafc;
-  padding: 11px;
-  border-radius: 8px;
-  font-weight: 600;
-  font-size: 12px;
-  cursor: pointer;
-}
-
-.tab.active {
-  background: #005aff;
-  color: white;
-  border-color: #005aff;
-}
-
-.leaflet-map-wrapper {
-  position: relative;
-  height: 390px;
-  border: 1px solid #dbe3ef;
-  border-radius: 12px;
-  overflow: hidden;
-  background: #eef3f9;
-}
-
-.leaflet-map {
-  width: 100%;
-  height: 100%;
-}
-
-.leaflet-control-attribution {
-  font-size: 9px;
-}
-
-.map-legend {
-  position: absolute;
-  left: 18px;
-  bottom: 18px;
-  z-index: 500;
-  background: white;
-  border-radius: 6px;
-  padding: 8px;
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  font-size: 11px;
-  box-shadow: 0 8px 20px rgba(15, 23, 42, 0.12);
-}
-
-.map-legend div {
-  width: 86px;
-  height: 10px;
-  background: linear-gradient(
-    90deg,
-    #ef4444,
-    #eab308,
-    #22c55e,
-    #064e3b
-  );
-}
-
-.kpi-card {
-  padding: 18px;
-}
-
-.kpi-title {
-  font-size: 18px;
-  font-weight: 700;
-  margin-bottom: 14px;
-}
-
-.kpi-content {
-  display: block;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 12px;
-}
-
-th {
-  text-align: left;
-  font-size: 11px;
-  color: #334155;
-  padding-bottom: 10px;
-}
-
-td {
-  padding: 8px 0;
-  font-weight: 500;
-  border-top: 1px solid #e5ebf3;
-}
-
-.last-grid {
-  display: grid;
-  grid-template-columns: 1.2fr repeat(4, 1fr);
-  gap: 12px;
-}
-
-.last-grid > div,
-.metric {
-  background: #f8fafc;
-  border: 1px solid #e5ebf3;
-  border-radius: 10px;
-  padding: 14px;
-}
-
-.session-info strong {
-  overflow-wrap: anywhere;
-}
-
-.session-info small {
-  display: block;
-  margin-top: 8px;
-  color: #64748b;
-  font-size: 11px;
-}
-
-span {
-  display: block;
-  color: #64748b;
-  font-size: 12px;
-}
-
-.metric strong,
-.last-grid strong {
-  display: block;
-  font-size: 22px;
-  color: #0f172a;
-  margin-top: 8px;
-}
-
-.phones-layout {
-  display: grid;
-  grid-template-columns: 280px 1fr;
-  gap: 24px;
-}
-
-.phone-select {
-  width: 100%;
-  text-align: left;
-  background: white;
-  border: 1px solid #e5ebf3;
-  border-radius: 10px;
-  padding: 16px;
-  margin-bottom: 12px;
-  cursor: pointer;
-  color: #0f172a;
-}
-
-.phone-select strong {
-  display: block;
-  margin-bottom: 6px;
-}
-
-.phone-select.active {
-  border-color: #005aff;
-  background: #f3f7ff;
-  color: #005aff;
-}
-
-.phone-title {
-  display: flex;
-  align-items: center;
-  margin-bottom: 18px;
-}
-
-.phone-title h2 {
-  margin: 0;
-}
-
-.phone-title p {
-  margin: 4px 0 0;
-  color: #64748b;
-  font-size: 13px;
-}
-
-.detail-tabs {
-  display: flex;
-  gap: 28px;
-  border-bottom: 1px solid #e5ebf3;
-  margin-bottom: 18px;
-}
-
-.detail-tabs button {
-  border: 0;
-  background: transparent;
-  padding: 0 0 12px;
-  cursor: pointer;
-}
-
-.detail-tabs .active {
-  color: #005aff;
-  border-bottom: 3px solid #005aff;
-}
-
-.summary-row {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  padding: 14px;
-  margin-bottom: 18px;
-}
-
-.recent-sessions {
-  margin-bottom: 18px;
-}
-
-.recent-sessions-grid {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  gap: 12px;
-}
-
-.recent-session-card {
-  border: 1px solid #e5ebf3;
-  border-radius: 10px;
-  padding: 14px;
-  background: #f8fafc;
-  text-align: center;
-  cursor: pointer;
-}
-
-.recent-session-card:disabled {
-  cursor: default;
-}
-
-.recent-session-card.active {
-  border-color: #005aff;
-  background: #f3f7ff;
-}
-
-.recent-session-card strong {
-  display: block;
-  font-size: 13px;
-  margin-bottom: 6px;
-}
-
-.phone-main-grid {
-  display: grid;
-  grid-template-columns: 1.15fr 1.15fr 1fr;
-  gap: 18px;
-}
-
-.phone-main-grid .leaflet-map-wrapper {
-  height: 300px;
-}
-
-.phone-kpi-panel {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-}
-
-.parameter-empty {
-  min-height: 300px;
-  display: flex;
-  flex-direction: column;
-}
-
-.parameter-empty p {
-  margin: auto;
-  color: #64748b;
-  font-size: 14px;
-  text-align: center;
-}
-
-.empty-box {
-  color: #64748b;
-  font-size: 14px;
-}
-
-@media (max-width: 1300px) {
-  .dashboard-grid,
-  .phone-main-grid,
-  .phones-layout {
-    grid-template-columns: 1fr;
-  }
-
-  .last-grid,
-  .summary-row,
-  .recent-sessions-grid {
-    grid-template-columns: 1fr 1fr;
-  }
-}
-
-.kpi-param-tabs {
-  display: flex;
-  gap: 8px;
-  margin-bottom: 14px;
-  flex-wrap: wrap;
-}
-
-.kpi-param {
-  border: 1px solid #dbe3ef;
-  background: #f8fafc;
-  border-radius: 8px;
-  padding: 8px 10px;
-  font-size: 11px;
-  cursor: pointer;
-}
-
-.kpi-param.active {
-  background: #005aff;
-  color: white;
-  border-color: #005aff;
-}
-
-.selected-row td {
-  color: #005aff;
-  font-weight: 700;
-}
-
-tbody tr {
-  cursor: pointer;
-}
-
-
-/*przycisk do propagacji*/
-.view-toggle {
-  display: flex;
-  gap: 8px;
-  margin-bottom: 12px;
-}
-
-.view-toggle-btn {
-  padding: 6px 16px;
-  border-radius: 20px;
-  cursor: pointer;
-  font-weight: 600;
-  font-size: 13px;
-  background: transparent;
-}
-
-.view-toggle-btn.measurements {
-  border: 2px solid #3b82f6;
-  color: #3b82f6;
-}
-
-.view-toggle-btn.measurements.active {
-  background: #3b82f6;
-  color: white;
-}
-
-.view-toggle-btn.propagation {
-  border: 2px solid #3b82f6;
-  color: #3b82f6;
-}
-
-.view-toggle-btn.propagation.active {
-  background: #3b82f6;
-  color: white;
 }
