+++ frontend/src/components/Heatmap.css
@@ -1,0 +1,106 @@
+.map-card {
+  padding: 18px;
+}
+
+.tabs {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 8px;
+  margin-bottom: 14px;
+}
+
+.tab {
+  border: 1px solid #dbe3ef;
+  background: #f8fafc;
+  padding: 11px;
+  border-radius: 8px;
+  font-weight: 600;
+  font-size: 12px;
+  cursor: pointer;
+}
+
+.tab.active {
+  background: #005aff;
+  color: white;
+  border-color: #005aff;
+}
+
+.leaflet-map-wrapper {
+  position: relative;
+  height: 390px;
+  border: 1px solid #dbe3ef;
+  border-radius: 12px;
+  overflow: hidden;
+  background: #eef3f9;
+}
+
+.leaflet-map {
+  width: 100%;
+  height: 100%;
+}
+
+.leaflet-control-attribution {
+  font-size: 9px;
+}
+
+.map-legend {
+  position: absolute;
+  left: 18px;
+  bottom: 18px;
+  z-index: 500;
+  background: white;
+  border-radius: 6px;
+  padding: 8px;
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  font-size: 11px;
+  box-shadow: 0 8px 20px rgba(15, 23, 42, 0.12);
+}
+
+.map-legend div {
+  width: 86px;
+  height: 10px;
+  background: linear-gradient(
+    90deg,
+    #ef4444,
+    #eab308,
+    #22c55e,
+    #064e3b
+  );
+}
+
+.view-toggle {
+  display: flex;
+  gap: 8px;
+  margin-bottom: 12px;
+}
+
+.view-toggle-btn {
+  padding: 6px 16px;
+  border-radius: 20px;
+  cursor: pointer;
+  font-weight: 600;
+  font-size: 13px;
+  background: transparent;
+}
+
+.view-toggle-btn.measurements {
+  border: 2px solid #3b82f6;
+  color: #3b82f6;
+}
+
+.view-toggle-btn.measurements.active {
+  background: #3b82f6;
+  color: white;
+}
+
+.view-toggle-btn.propagation {
+  border: 2px solid #3b82f6;
+  color: #3b82f6;
+}
+
+.view-toggle-btn.propagation.active {
+  background: #3b82f6;
+  color: white;
+}
