+++ frontend/src/components/Metric.css
@@ -1,0 +1,13 @@
+.metric {
+  background: #f8fafc;
+  border: 1px solid #e5ebf3;
+  border-radius: 10px;
+  padding: 14px;
+}
+
+.metric strong {
+  display: block;
+  font-size: 22px;
+  color: #0f172a;
+  margin-top: 8px;
+}
