+++ frontend/src/components/PhoneList.css
@@ -1,0 +1,32 @@
+.phone-list {
+  padding: 18px;
+}
+
+.phone-select {
+  width: 100%;
+  text-align: left;
+  background: white;
+  border: 1px solid #e5ebf3;
+  border-radius: 10px;
+  padding: 16px;
+  margin-bottom: 12px;
+  cursor: pointer;
+  color: #0f172a;
+}
+
+.phone-select strong {
+  display: block;
+  margin-bottom: 6px;
+}
+
+.phone-select.active {
+  border-color: #005aff;
+  background: #f3f7ff;
+  color: #005aff;
+}
+
+.empty-box {
+  padding: 18px;
+  color: #64748b;
+  font-size: 14px;
+}
