+++ frontend/src/components/RecentSessions.css
@@ -1,0 +1,34 @@
+.recent-sessions {
+  padding: 18px;
+  margin-bottom: 18px;
+}
+
+.recent-sessions-grid {
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
+  gap: 12px;
+}
+
+.recent-session-card {
+  border: 1px solid #e5ebf3;
+  border-radius: 10px;
+  padding: 14px;
+  background: #f8fafc;
+  text-align: center;
+  cursor: pointer;
+}
+
+.recent-session-card:disabled {
+  cursor: default;
+}
+
+.recent-session-card.active {
+  border-color: #005aff;
+  background: #f3f7ff;
+}
+
+.recent-session-card strong {
+  display: block;
+  font-size: 13px;
+  margin-bottom: 6px;
+}
