+++ frontend/src/components/Sidebar.css
@@ -1,0 +1,45 @@
+.sidebar {
+  width: 180px;
+  min-height: 100vh;
+  background: #15224a;
+  color: white;
+  padding: 28px 18px;
+  display: flex;
+  flex-direction: column;
+  position: sticky;
+  top: 0;
+}
+
+.logo-container {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  padding-top: 28px;
+  margin-bottom: 40px;
+}
+
+.logo-image {
+  max-width: 170px;
+  height: auto;
+  object-fit: contain;
+  display: block;
+}
+
+.nav {
+  width: 100%;
+  border: 0;
+  color: white;
+  background: transparent;
+  padding: 14px 16px;
+  border-radius: 10px;
+  text-align: left;
+  cursor: pointer;
+  margin-bottom: 8px;
+  font-size: 14px;
+}
+
+.nav.active,
+.nav:hover {
+  background: #005aff;
+}
