+++ frontend/src/pages/Dashboard.css
@@ -1,0 +1,34 @@
+.page {
+  flex: 1;
+  padding: 28px 32px;
+}
+
+.topbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 24px;
+}
+
+.topbar h1 {
+  margin: 0;
+  font-size: 28px;
+}
+
+.dashboard-grid {
+  display: grid;
+  grid-template-columns: 1.15fr 1.35fr;
+  gap: 24px;
+  margin-bottom: 24px;
+}
+
+.dashboard-map-panel {
+  padding: 18px;
+}
+
+.dashboard-kpi-panel {
+  padding: 18px;
+  display: flex;
+  flex-direction: column;
+  gap: 18px;
+}
