+++ frontend/src/styles/global.css
@@ -1,0 +1,51 @@
+* {
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  font-family: "Poppins", Arial, sans-serif;
+  background: #f8fafc;
+  color: #0f172a;
+}
+
+button {
+  font-family: inherit;
+}
+
+span {
+  display: block;
+  color: #64748b;
+  font-size: 12px;
+}
+
+h2,
+h3 {
+  margin: 0 0 12px;
+}
+
+table {
+  width: 100%;
+  border-collapse: collapse;
+  font-size: 12px;
+}
+
+th {
+  text-align: left;
+  font-size: 11px;
+  color: #334155;
+  padding-bottom: 10px;
+}
+
+td {
+  padding: 8px 0;
+  font-weight: 500;
+  border-top: 1px solid #e5ebf3;
+}
+
+.card {
+  background: white;
+  border: 1px solid #dbe3ef;
+  border-radius: 14px;
+  box-shadow: 0 12px 30px rgba(15, 23, 42, 0.04);
+}
